feat(fields): type and name order changed in all Field classes
Type got higher priority over name when constructing Fields. Name is not mandatory anymore, random
string is generated if empty string is passed.

BREAKING CHANGE: Name is not mandatory anymore for Field objects. Also it swapped place with type.
This is done not only for logical reasons, but also that nested classes, like List, Dict and similar
does not have to provide name during object creation.